Rename expandHoverTop and expandHoverBottom.
