+++ src/reducers/index.ts
@@ -29,8 +29,8 @@ export { default as editableRender } from './editableRender'
 export { default as editing } from './editing'
 export { default as editThought } from './editThought'
 export { default as error } from './error'
-export { default as expandBottom } from './expandBottom'
 export { default as expandContextThought } from './expandContextThought'
+export { default as expandHoverBottom } from './expandHoverBottom'
 export { default as expandHoverTop } from './expandHoverTop'
 export { default as extractThought } from './extractThought'
 export { default as fontSize } from './fontSize'
